Price changed from integer to floats

+++ data/migrations/20190413152117_books.js
@@ -24,7 +24,7 @@ exports.up = function(knex, Promise) {
         .onUpdate("CASCADE");
       table.string("author", 40).notNullable();
       table.string("name", 40).notNullable();
-      table.integer("price").notNullable();
+      table.float("price").notNullable();
       table.string("publisher", 40).notNullable();
       table.string("description", 600);
       table.string("imageUrl", 256);
